Search History Support

+++ Delete.jsp
@@ -4,8 +4,10 @@
 File directory = new File("database");
 File[] files = directory.listFiles();
 for (File file : files)
-	if(file.delete())
-		out.println("Sucessfully deleted file: "+file.getName());
-	else
-		out.println("Error in deleting file");
+	if(file.getName().compareTo("SearchHistory.db") != 0 && 
+		file.getName().compareTo("SearchHistory.lg") != 0)
+		if(file.delete())
+			out.println("Sucessfully deleted file: "+file.getName());
+		else
+			out.println("Error in deleting file");
 %>
